+++ __tests__/handlers.test.js
@@ -1,12 +1,23 @@
-const util = require("../lib/utils");
 const geekHandler = require("../lib/handlers");
+jest.mock('../lib/handlers');
 
 test('GeekExecuteCommand Function - execute', (done) => {
-    function callback(stdout) {
-        expect(stdout).not.toBeNull();
-        done();
-    }
 
-    geekHandler.geekExecuteCommand('ls -a')().then(callback)
+    const resp = 'some';
+    geekHandler.geekExecuteCommand.mockResolvedValue(resp);
+
+    geekHandler.geekExecuteCommand('ls -a').then((res) => {
+        expect(res).not.toBeNull();
+        done();
+    })
 });
 
+test('CopyJSONFile Function - execute', (done) => {
+    const resp = true;
+    geekHandler.copyJsonFile.mockResolvedValue(resp);
+
+    geekHandler.copyJsonFile('copy.me', 'paste.me').then((res => {
+        expect(res).toBe(true);
+        done();
+    }))
+});
