+++ lib/handlers.js
@@ -29,6 +29,18 @@ let geekExecuteCommand = (command = 'ls') => {
 	}
 }
 
+let copyJsonFile = (source, target) => {
+	return () => {
+		return new Promise(function (resolve, reject) {
+			fs.copyFile(source, target, (err) => {
+				if (err) reject(err);
+				else {
+					resolve(true)
+				}
+			});
+		});
+	}
+}
 
 let readJsonFile = (fileName) => {
 	return () => {
@@ -113,7 +125,7 @@ let geekInit = () => {
 		printTitle(),
 		geekExecuteCommand('git fetch origin'),
 		geekExecuteCommand('git checkout geek-config'),
-		utils.copyJsonFile(path.join(process.cwd(), "geek-config.json"), path.join(process.cwd(), "geek.json")),
+		copyJsonFile(path.join(process.cwd(), "geek-config.json"), path.join(process.cwd(), "geek.json")),
 		geekExecuteCommand('git checkout master'),
 	]).then((res) => {
 		console.log('Success')
@@ -260,6 +272,7 @@ let geekDiff = () => {
 }
 
 module.exports = {
+	"copyJsonFile":copyJsonFile,
 	"geekPerformInitAction": geekInit,
 	"geekPerformResetAction": geekReset,
 	"geekPerformNextAction": geekNext,
