+++ lib/utils.js
@@ -27,22 +27,7 @@ const PromiseSeries = (arr = []) => {
 	}, ret)
 }
 
-let copyJsonFile = (source, target) => {
-	return () => {
-		return new Promise(function (resolve, reject) {
-			fs.copyFile(source, target, (err) => {
-				if (err) reject(err);
-				else {
-					resolve(true)
-				}
-			});
-		});
-	}
-}
-
-
 module.exports = {
 	execute,
-	copyJsonFile,
 	PromiseSeries
 }
